+++ KkuMulKum/Source/Promise/ReadyStatus/View/SetReadyInfoView.swift
