+++ KkuMulKum/Source/Promise/ReadyStatus/ViewModel/SetReadyInfoViewModel.swift
@@ -1,0 +1,9 @@
+//
+//  SetReadyInfoViewModel.swift
+//  KkuMulKum
+//
+//  Created by 예삐 on 7/15/24.
+//
+
+import UIKit
+
